[ANDWL-239] refactoring main menu
